+++ source/extensions/omni.isaac.benchmarks/omni/isaac/benchmarks/tests/test_benchmark_robots_o3dyn.py
@@ -35,9 +35,8 @@ class TestBenchmarkRobotsO3dyn(BaseIsaacBenchmarkAsync):
         #     sensor_name += "_lidar"
         if len(sensor_name) == 0:
             sensor_name = "_no_sensor"
-        self.test_run.test_name = f"robots_o3dyn{sensor_name}_{n_robot}"
+        self.benchmark_name = f"robots_o3dyn{sensor_name}_{n_robot}"
         self.set_phase("loading")
-        self.start_runtime()
 
         robot_path = "/Isaac/Robots/O3dyn/o3dyn.usd"
         scene_path = "/Isaac/Environments/Simple_Warehouse/full_warehouse.usd"
@@ -93,17 +92,14 @@ class TestBenchmarkRobotsO3dyn(BaseIsaacBenchmarkAsync):
         await omni.kit.app.get_app().next_update_async()
         await omni.kit.app.get_app().next_update_async()
 
-        self.stop_runtime()
         await self.store_measurements()
 
         # perform benchmark
         self.set_phase("benchmark")
-        self.start_collecting_frametime()
 
         for _ in range(1 if self.test_mode else TEST_NUM_APP_UPDATES):
             await omni.kit.app.get_app().next_update_async()
 
-        self.stop_collecting_frametime()
         await self.store_measurements()
 
         timeline.stop()