+++ source/extensions/omni.isaac.benchmarks/omni/isaac/benchmarks/tests/test_benchmark_rtx_radar.py
@@ -29,9 +29,8 @@ class TestBenchmarkRtxRadar(BaseIsaacBenchmarkAsync):
 
     # ----------------------------------------------------------------------
     async def benchmark_rtx_radar(self, n_sensor):
-        self.test_run.test_name = f"rtx_radar_{n_sensor}"
+        self.benchmark_name = f"rtx_radar_{n_sensor}"
         self.set_phase("loading")
-        self.start_runtime()
         scene_path = "/Isaac/Environments/Simple_Warehouse/full_warehouse.usd"
         await self.fully_load_stage(self.assets_root_path + scene_path)
         timeline = omni.timeline.get_timeline_interface()
@@ -64,17 +63,14 @@ class TestBenchmarkRtxRadar(BaseIsaacBenchmarkAsync):
 
             await omni.kit.app.get_app().next_update_async()
 
-        self.stop_runtime()
         await self.store_measurements()
 
         self.set_phase("benchmark")
         timeline.play()
-        self.start_collecting_frametime()
 
         for _ in range(1 if self.test_mode else TEST_NUM_APP_UPDATES):
             await omni.kit.app.get_app().next_update_async()
 
-        self.stop_collecting_frametime()
         await self.store_measurements()
 
         timeline.stop()