+++ source/extensions/omni.isaac.benchmarks/omni/isaac/benchmarks/tests/test_benchmark_sdg_generation.py
@@ -83,8 +83,8 @@ class TestBenchmarkSDGGeneration(BaseIsaacBenchmarkAsync):
     # ----------------------------------------------------------------------
     async def benchmark_sdg_generation(self, num_render_products, resolution, num_frames, annotators="rgb"):
         # Set the test name
-        self.test_run.test_name = f"sdg_generation_cameras_{num_render_products}_resolution_{resolution[0]}_{resolution[1]}_frames_{num_frames}_{annotators}_annotators"
-        print(f" ** [sdg_benchmark] Running test: {self.test_run.test_name}")
+        self.benchmark_name = f"sdg_generation_cameras_{num_render_products}_resolution_{resolution[0]}_{resolution[1]}_frames_{num_frames}_{annotators}_annotators"
+        print(f" ** [sdg_benchmark] Running test: {self.benchmark_name}")
 
         # Create a fresh stage
         await create_new_stage_async()
@@ -95,8 +95,6 @@ class TestBenchmarkSDGGeneration(BaseIsaacBenchmarkAsync):
         # Perform the loading phase: check for the duration for loading the stage,
         # where the first frames take longer in order to load all the prims followed by the materials/textures
         self.set_phase("loading")
-        # self.start_collecting_frametime()
-        self.start_runtime()
         await omni.kit.app.get_app().next_update_async()
 
         # Async version seems to cause running extra frames when collecting frametimes
@@ -105,21 +103,17 @@ class TestBenchmarkSDGGeneration(BaseIsaacBenchmarkAsync):
         await wait_until_stage_is_fully_loaded_async()
 
         await omni.kit.app.get_app().next_update_async()
-        self.stop_runtime()
-        # self.stop_collecting_frametime()
         await self.store_measurements()
         # ----------------------------- loading phase ----------------------------->
 
         # <---------------------------- baseline phase -----------------------------
         # Perform the baseline phase: check the performance with the cleanly loaded stage
         self.set_phase("baseline")
-        self.start_collecting_frametime()
 
         # Run for a few frames to check the clean stage stats
         for _ in range(1 if self.test_mode else TEST_NUM_APP_UPDATES):
             await omni.kit.app.get_app().next_update_async()
 
-        self.stop_collecting_frametime()
         await self.store_measurements()
         # ---------------------------- baseline phase ------------------------------->
 
@@ -130,7 +124,7 @@ class TestBenchmarkSDGGeneration(BaseIsaacBenchmarkAsync):
             cam = rep.create.camera(position=(-6, -1, 4 + z_offset), look_at=LOOK_AT_PRIM_PATH, name=f"cam_{i}")
             render_products.append(rep.create.render_product(cam, resolution))
         writer = rep.WriterRegistry.get("BasicWriter")
-        output_directory = os.getcwd() + "/" + self.test_run.test_name
+        output_directory = os.getcwd() + "/" + self.benchmark_name
         if annotators == "rgb":
             writer.initialize(output_dir=output_directory, rgb=True)
         elif annotators == "all":
@@ -142,13 +136,11 @@ class TestBenchmarkSDGGeneration(BaseIsaacBenchmarkAsync):
         # <---------------------------- baseline_sdg phase -----------------------------
         # Check the performance with the SDG parts loaded
         self.set_phase("baseline_sdg")
-        self.start_collecting_frametime()
 
         # Run for a few frames to check the stage stats with the render products
         for _ in range(1 if self.test_mode else TEST_NUM_APP_UPDATES):
             await omni.kit.app.get_app().next_update_async()
 
-        self.stop_collecting_frametime()
         await self.store_measurements()
         # ---------------------------- baseline_sdg phase ------------------------------->
 
@@ -159,8 +151,6 @@ class TestBenchmarkSDGGeneration(BaseIsaacBenchmarkAsync):
         # <---------------------------- benchmark phase -----------------------------
         # Check the performance while writing SDG data
         self.set_phase("benchmark")
-        self.start_collecting_frametime()
-        self.start_runtime()
 
         await rep.orchestrator.run_until_complete_async(num_frames=(1 if self.test_mode else num_frames))
         # await rep.orchestrator.run_async(num_frames=num_frames)
@@ -168,8 +158,6 @@ class TestBenchmarkSDGGeneration(BaseIsaacBenchmarkAsync):
         #     await rep.orchestrator.step_async()
         # await rep.orchestrator.wait_until_complete_async()
 
-        self.stop_runtime()
-        self.stop_collecting_frametime()
         await self.store_measurements()
         # ---------------------------- benchmark phase ------------------------------->
 
@@ -196,13 +184,11 @@ class TestBenchmarkSDGGeneration(BaseIsaacBenchmarkAsync):
         # <---------------------------- baseline_cleanup phase -----------------------------
         # Check performance after running the SDG benchmark and cleaning up replicator functionalities
         self.set_phase("baseline_cleanup")
-        self.start_collecting_frametime()
 
         # Run for a few frames to check the stage stats after detaching the writer and destroying the render products
         for _ in range(1 if self.test_mode else TEST_NUM_APP_UPDATES):
             await omni.kit.app.get_app().next_update_async()
 
-        self.stop_collecting_frametime()
         await self.store_measurements()
         # ---------------------------- baseline_cleanup phase ------------------------------->