+++ source/standalone_examples/benchmarks/benchmark_camera.py
@@ -40,10 +40,18 @@ from omni.isaac.benchmark.services import base_isaac_benchmark
 
 # Create the benchmark
 benchmark = base_isaac_benchmark.BaseIsaacBenchmark(
-    benchmark_name=f"cameras_{n_camera}_resolution_{resolution[0]}_{resolution[1]}_gpu_{n_gpu}"
+    benchmark_name="benchmark_camera",
+    workflow_metadata={
+        "metadata": [
+            {"name": "num_cameras", "data": n_camera},
+            {"name": "width", "data": resolution[0]},
+            {"name": "height", "data": resolution[1]},
+            {"name": "num_gpus", "data": n_gpu},
+        ]
+    },
 )
 benchmark.set_phase("loading")
-benchmark.start_runtime()
+
 
 scene_path = "/Isaac/Environments/Simple_Warehouse/full_warehouse.usd"
 benchmark.fully_load_stage(benchmark.assets_root_path + scene_path)
@@ -76,17 +84,13 @@ while is_stage_loading():
     omni.kit.app.get_app().update()
 omni.kit.app.get_app().update()
 
-benchmark.stop_runtime()
 benchmark.store_measurements()
-
 # perform benchmark
 benchmark.set_phase("benchmark")
-benchmark.start_collecting_frametime()
 
 for _ in range(1 if benchmark.test_mode else TEST_NUM_APP_UPDATES):
     omni.kit.app.get_app().update()
 
-benchmark.stop_collecting_frametime()
 benchmark.store_measurements()
 benchmark.stop()